State: Move user-profile-links to profile-links
